adds path to output files
It seems that that the current bwa-mem2 index wrapper script does not respect paths in the ouput files. this leads to an error in the workflow execution as the expected output files are not written to the correct path, but rather to the root directory from which the Snakefile is executed.